improvements in 'injest' performance.  Change from individual DB inserts to using bulk_create.  this had two major reprocuions, 1. SQLite doesn't handle it very well due to it can only have 1000 ops per sql statement, so 2. had to switch to using a local windows copy of Postgres.  There were some fields that needed coaxing, but all in all pretty straight forward.
Using postgres turns out to be a pretty good option, in that it will move
better to a hosted environment.

+++ manager/models.py
@@ -104,9 +104,9 @@ class DiskSpace(models.Model):
         diskfree  - How much free space is there.
         date      - When was this entry made (self updating).
     """
-    cloud = models.IntegerField()
-    size  = models.IntegerField()
-    free  = models.IntegerField()
+    cloud = models.BigIntegerField()
+    size  = models.BigIntegerField()
+    free  = models.BigIntegerField()
     date  = models.DateTimeField()
     
     client = models.ForeignKey(ManagedClient)
